Refactor string extensions and update method signatures
Refactor to organize string-related extension methods into the
`MG.Sonarr.Next.Extensions.Strings` namespace across multiple files.
Removed guard clauses and updated method signatures in
`SpanExtensions.cs` to use `ReadOnlySpan<char>`. Introduced a new
private method `FirstCharEquals` for character comparison. Removed
several methods in `StringExtensions.cs` related to string splitting
and hash code generation. Updated `StartsWith` method in
`JsonModifiers.cs` to remove the `in` keyword from the `underscore`
parameter.
