+++ src/MG.Sonarr.Next.Shell/Attributes/ValidateUrlAttribute.cs
@@ -1,4 +1,5 @@
 ﻿using MG.Sonarr.Next.Extensions;
+using MG.Sonarr.Next.Extensions.Strings;
 
 namespace MG.Sonarr.Next.Shell.Attributes
 {
