+++ src/MG.Sonarr.Next/Extensions/EnumExtensions.cs
@@ -1,4 +1,6 @@
-﻿namespace MG.Sonarr.Next.Extensions
+﻿using MG.Sonarr.Next.Extensions.Strings;
+
+namespace MG.Sonarr.Next.Extensions
 {
     /// <summary>
     /// Custom extension methods for <see cref="Enum"/> classes.
