+++ src/MG.Sonarr.Next/Extensions/HttpStatusCodeExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using MG.Sonarr.Next.Attributes;
+using MG.Sonarr.Next.Extensions.Strings;
 using System.Collections.ObjectModel;
 using System.Net;
