+++ src/MG.Sonarr.Next/Extensions/Strings/StringExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using MG.Sonarr.Next.Attributes;
 
-namespace MG.Sonarr.Next.Extensions
+namespace MG.Sonarr.Next.Extensions.Strings
 {
     /// <summary>
     /// Custom extension methods for high-performance split operations of <see cref="string"/> 
@@ -8,37 +8,6 @@ namespace MG.Sonarr.Next.Extensions
     /// </summary>
     public static class SplittingExtensions
     {
-        
-        public static int GetIgnoreCaseHashCode(this string str)
-        {
-            ArgumentNullException.ThrowIfNull(str);
-            return StringComparer.InvariantCultureIgnoreCase.GetHashCode(str);
-        }
-
-        /// <summary>
-        /// A zero-allocation method for enumerating a <see cref="ReadOnlySpan{T}"/> source into sections
-        /// split by a given character.
-        /// </summary>
-        /// <remarks>
-        ///     As <see cref="SplitEnumerator"/> is a ref struct, this method cannot be used in
-        ///     <see langword="async"/> methods.
-        ///     <para>
-        ///         An example of utilizing this method looks like:
-        ///         <code>foreach (ReadOnlySpan&lt;char&gt; section in "key:value".AsSpan().SpanSplit(':'))</code>
-        ///     </para>
-        /// </remarks>
-        /// <param name="value">The source span to split.</param>
-        /// <param name="splitBy">The character value to split the source span on.</param>
-        /// <returns>
-        ///     A ref struct enumerator that loops through each <see cref="ReadOnlySpan{T}"/> section when
-        ///     split by <paramref name="splitBy"/>.
-        /// </returns>
-        [DebuggerStepThrough]
-        public static SplitEnumerator SpanSplit(this ReadOnlySpan<char> value, in char splitBy)
-        {
-            Guard.IsSpan(value);
-            return SpanSplit(value, new ReadOnlySpan<char>(in splitBy));
-        }
         /// <summary>
         /// A zero-allocation method for enumerating a <see cref="ReadOnlySpan{T}"/> source into sections
         /// split by a given character.
