+++ src/MG.Sonarr.Next/Models/System/PingResult.cs
@@ -1,4 +1,4 @@
-﻿using MG.Sonarr.Next.Extensions;
+﻿using MG.Sonarr.Next.Extensions.Strings;
 using MG.Sonarr.Next.Services.Http;
 using System.Management.Automation;
 using System.Net;
