+++ src/MG.Sonarr.Next/PSProperties/ReadOnlyProperty.cs
@@ -1,4 +1,5 @@
 ﻿using MG.Sonarr.Next.Extensions;
+using MG.Sonarr.Next.Extensions.Strings;
 using System.Management.Automation;
 
 namespace MG.Sonarr.Next.PSProperties
